changing congif server access. edits to all files to clean up unused code.
