--- views/ghosts.handlebars
@@ -1,11 +1,0 @@
-{{!-- <h1 id="id"></h1>
-<script>
-    $.ajax({
-    method: "get",
-    url: "/api/ghosts/1",
-    }).then(function(data) {
-       document.querySelector("#id").textContent = data.id
-    })
-</script> --}}
-{{!-- <h1>{{fullName}}</h1>
-<h1>ghost page</h1> --}}
