+++ views/home.handlebars
@@ -127,7 +127,7 @@
 
 							<div class="form-group">
 								<label class="label-text" for="age">What is your age?</label>
-								<input id="age" type="number" min="1" max="100" value="78">
+								<input id="age" type="number" min="1" max="100" value="30">
 							</div>
 							<div class="form-group">
 								<label class="label-text" for="gender">What is your gender?</label>
