+++ views/index.handlebars
@@ -17,14 +17,11 @@
 				<div class="cell small-5 large-5"></div>
 				<div class="center cell small-2 large-2">
 
-					{{!-- <p class="dialogue">{{insert current line of text here}}</p>
-					<p>{{insert next dialogue line here}}</p> --}}
 					<a class="button large secondary" id="newStartEnterBtn" href="/startpage/1">An eerie feeling compels you towards the house!</a>
 					<br>
 				</div>
 				<div class="cell small-5 large-5">
-					{{!-- <button id="positive">{{positive-response}}</button>
-					<button id="negative">{{negative-response}}</button> --}}
+					
 				</div>
 			</div>
 		</container>
