+++ views/layouts/main.handlebars
@@ -20,9 +20,6 @@
 
 <body>
 	{{{ body }}}
-	    {{!-- <script src="js/vendor/jquery.js"></script>
-    <script src="js/vendor/what-input.js"></script>
-    <script src="js/vendor/foundation.min.js"></script> --}}
     <script>
       $(document).foundation();
     </script>
