+++ views/room1.handlebars
@@ -10,7 +10,6 @@
         </div>
     </div>
 </div>
-{{!-- <img src="#" id="ghostSprite" alt="ghost_sprite" /> --}}
 <div class="footer">
     <hr>
     <div class="grid-y medium-grid-frame">
@@ -241,8 +240,6 @@
 
                 </div>
                 <div class="medium-5 column">
-
-
                 </div>
                 <div class="medium-1 column"></div>
             </div>
