+++ views/room2.handlebars
@@ -1,11 +1,8 @@
-
 <div class="grid-x">
     <div class="cell">
         <div class="center">
-            <img src="https://ashlynimages.s3.us-east-2.amazonaws.com/Attic.jpg" class="bg3"
-                alt="Room 2 Image" />
-            <img id="room2GhostSprite"
-                src="https://ashlynimages.s3.us-east-2.amazonaws.com/minerGhostPXL.png"
+            <img src="https://ashlynimages.s3.us-east-2.amazonaws.com/Attic.jpg" class="bg3" alt="Room 2 Image" />
+            <img id="room2GhostSprite" src="https://ashlynimages.s3.us-east-2.amazonaws.com/minerGhostPXL.png"
                 alt="ghost" />
         </div>
     </div>
@@ -53,7 +50,8 @@
                             <p>
                                 `Hahaha. I'm just messing with you. I get bored sometimes and we don't see many
                                 newcomers 'round
-                                here. My name is {{entryGhost.fullName}}, and I travelled this whole country looking for ore
+                                here. My name is {{entryGhost.fullName}}, and I travelled this whole country looking for
+                                ore
                                 before the
                                 railroads were
                                 even built.`
@@ -84,7 +82,8 @@
                                 `Hah. Get yourself together. Why are you so scared of everything? I travelled this whole
                                 land
                                 barefoot,
-                                and I remember everyone being scared of Ol' {{entryGhost.fullName}}, not the other way around!`
+                                and I remember everyone being scared of Ol' {{entryGhost.fullName}}, not the other way
+                                around!`
                             </p>
 
                         </div>
@@ -111,10 +110,12 @@
                             <p>
                                 `AHAHAHA I spent 18 hours a day digging through the dirt and you ask why I smell bad? Of
                                 course
-                                I do! I came from {{entryGhost.homeTown}}, but I dug hard and I even owned my own mine for
+                                I do! I came from {{entryGhost.homeTown}}, but I dug hard and I even owned my own mine
+                                for
                                 awhile. The people
                                 around here
-                                called it Ol' {{entryGhost.fullName}}'s Mud Pit. I think we might be on top of it now. This
+                                called it Ol' {{entryGhost.fullName}}'s Mud Pit. I think we might be on top of it now.
+                                This
                                 house was never
                                 here for me
                                 to see it before I just kind of...showed up.
@@ -142,9 +143,11 @@
                         <div class="form-group">
 
                             <p>
-                                `Of course I did! What else would I have dug for? I left {{entryGhost.homeTown}} to dig for stuff
+                                `Of course I did! What else would I have dug for? I left {{entryGhost.homeTown}} to dig
+                                for stuff
                                 when I was
-                                {{entryGhost.deadFor}} years old. Years later, there was an accident, I think, and all of the sudden
+                                {{entryGhost.deadFor}} years old. Years later, there was an accident, I think, and all
+                                of the sudden
                                 I'm stuck
                                 here in this
                                 fancy house.`
@@ -171,7 +174,8 @@
                         <div class="form-group">
 
                             <p>
-                                Will you help me, friend? I think I'm trapped, and as much I think {{frontGhost.fullName}} is a
+                                Will you help me, friend? I think I'm trapped, and as much I think
+                                {{frontGhost.fullName}} is a
                                 nice
                                 lady, this house isn't my own and I need to keep searching for gold.
                             </p>
@@ -219,17 +223,12 @@
 
                     <a class="button secondary run" id="runBtn" href="/foyer/1">RUN!</a>
 
-
                 </div>
                 <div class="medium-5 column">
-
-
                 </div>
                 <div class="medium-1 column"></div>
             </div>
         </div>
-      </div>
-         
-            <div class="medium-1 column"></div>
-          
     </div>
+    <div class="medium-1 column"></div>
+</div>
