+++ views/room3.handlebars
@@ -1,4 +1,3 @@
-
 <div class="grid-x">
     <div class="cell">
         <div class="center">
