+++ views/startpage.handlebars
@@ -16,27 +16,12 @@
 			<div class="grid-x">
 				<div class="cell small-5 large-5"></div>
 				<div class="center cell small-2 large-2">
-
-					{{!-- <p class="dialogue">{{insert current line of text here}}</p>
-					<p>{{insert next dialogue line here}}</p> --}}
 					<a class="button large secondary" id="startBtn" href="/home/1">See what is inside!</a>
 					<br>
 				</div>
 				<div class="cell small-5 large-5">
-					{{!-- <button id="positive">{{positive-response}}</button>
-					<button id="negative">{{negative-response}}</button> --}}
 				</div>
 			</div>
 		</container>
 	</div>
 </main>
-{{!-- <script type= "text/javascript">
-		$.ajax({
-			method: "DELETE",
-			url: "/api/ghosts",
-		});
-		$.ajax({
-			method: "DELETE",
-			url: "/api/users",
-		});
-</script> --}}
